+++ src/main/java/com/example/electriccomponentsshop/dto/ProductDTO.java
@@ -1,11 +1,12 @@
 package com.example.electriccomponentsshop.dto;
 
+import com.example.electriccomponentsshop.entities.SpecificationValue;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-
+import javax.persistence.Column;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
@@ -18,7 +19,7 @@ public class ProductDTO {
     private String id;
     private String name;
     private String image;
-    private Integer available;
+    private BigInteger available;
     private BigDecimal price;
     private List<CategoryDTO> categories;
     private List<SpecificationValueDto> specificationValues;
